Po testach pierwszych budynków

+++ NewWorld/Models/Buildings.cs
@@ -11,7 +11,7 @@ namespace NewWorld.Models
         [Key]
         public int Id { get; set; }
         public int RezydencjaFarmerow { get; set; }
-        public int Farmerzy { get; set; }
+        public double Farmerzy { get; set; }
         public virtual Resources FarmersSatisfaction { get; set; }
         public int ChatkaRybacka { get; set; }
         //public virtual Island Island { get; set; }
